+++ backend/src/baserow/contrib/database/api/views/errors.py
@@ -61,3 +61,8 @@ ERROR_VIEW_DOES_NOT_SUPPORT_FIELD_OPTIONS = (
     HTTP_400_BAD_REQUEST,
     "This view model does not support field options.",
 )
+ERROR_CANNOT_SHARE_VIEW_TYPE = (
+    "ERROR_CANNOT_SHARE_VIEW_TYPE",
+    HTTP_400_BAD_REQUEST,
+    "This view type does not support sharing.",
+)